Crate ibc_core_commitment_types

ICS-23: Commitment implementation of a cryptographic scheme that verifies state transitions between chains.

Modules

commitment

Defines core commitment types

error

Defines the commitment error type

merkle

Merkle proof utilities

proto

Re-exports ICS-23 proto types from the ibc-proto crate, which are used in the implementation of dependent IBC crates.

serializer

specs

Defines proof specs, which encode the structure of proofs